+++ delta_utils/__init__.py
@@ -1,5 +1,6 @@
 from .core import (
     NoNewDataException,
+    ReadChangeFeedDisabled,
     last_written_timestamp_for_delta_path,
     read_change_feed,
 )
@@ -9,6 +10,7 @@ __all__ = (
     "read_change_feed",
     "last_written_timestamp_for_delta_path",
     "NoNewDataException",
+    "ReadChangeFeedDisabled",
     "DeltaChanges",
     "NonDeltaLastWrittenTimestamp",
     "new_and_updated",